Restructure functionality regarding high-level model checking, add more support for unsanitized mc
