+++ src/analysis.rs
@@ -2,7 +2,7 @@
 
 use crate::evaluation::algorithm::{compute_steady_states, eval_node};
 use crate::evaluation::eval_info::EvalInfo;
-use crate::model_checking::{collect_unique_hctl_vars, get_extended_symbolic_graph};
+use crate::mc_utils::{collect_unique_hctl_vars, get_extended_symbolic_graph};
 use crate::preprocessing::parser::parse_hctl_formula;
 use crate::preprocessing::utils::check_props_and_rename_vars;
 use crate::result_print::*;
