+++ src/lib.rs
@@ -8,6 +8,7 @@
 
 pub mod analysis;
 pub mod evaluation;
+pub mod mc_utils;
 pub mod model_checking;
 pub mod preprocessing;
 pub mod result_print;
